change App component directory
