+++ src/index.js
@@ -1,18 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import WelcomeScreen from './components/welcome-screen/welcome-screen.jsx';
-import PropTypes from 'prop-types';
-
-const App = (props) => {
-  const {gameTime, mistakesCount} = props;
-
-  return <WelcomeScreen gameTime={gameTime} mistakesCount={mistakesCount} />;
-};
-
-App.propTypes = {
-  gameTime: PropTypes.number,
-  mistakesCount: PropTypes.number,
-};
+import App from './components/app/app.jsx';
 
 const init = () => {
   const configs = {
